S3:EOS-16042: S3background delete Producer porting to Message Bus
    1. Code restructured into a single file.
